Fixes Issue #506, refactors `CbcParser`
- `CbcParser` --> `CbcReader` to follow in the conventions established by the
  `Recipe*` classes
- `CbcParser` still exists for compatibility purposes AND for future potential
  feature development
- Moves `CbcOutputType` and `GeneratedVariantsType` under `parser.types`
  instead of the `CbcParser` class.
- Improves documentation accordingly
